refactor: 타입 파일 분리(#61)
